otf2bfd.c: fix exit code when generate_font is run
The generate_font() function uses the result of fprintf() as its
return value. This is the number of chars written or a negative
number in case of an error. The otfbdf.c code uses that return
code as the exit-code of the entire program. This will break
scripts that use "set -e" because a successful run of otf2bdf
will result in an exit-code of 8 (`fprintf(out, "ENDFONT\n");`
is the last write).

This patch changes the return code of generate_font() to 0
in case there is no error and to -1 if there is an error.

This will lead to an exit-code of zero if generate_font()
was successful.